Struct ark_ed_on_bls12_381_bandersnatch::BandersnatchConfig

pub struct BandersnatchConfig;

bandersnatch is an incomplete twisted Edwards curve. These curves have equations of the form: ax² + y² = 1 + dx²y². over some base finite field Fq.

bandersnatch’s curve equation: -5x² + y² = 1 + dx²y²

q = 52435875175126190479447740508185965837690552500527637822603658699938581184513.

a = -5. d = (138827208126141220649022263972958607803/ 171449701953573178309673572579671231137) mod q = 45022363124591815672509500913686876175488063829319466900776701791074614335719.

Sage script to calculate these:

q = 52435875175126190479447740508185965837690552500527637822603658699938581184513
Fq = GF(q)
d = (Fq(138827208126141220649022263972958607803)/Fq(171449701953573178309673572579671231137))

These parameters and the sage script obtained from: https://github.com/asanso/Bandersnatch/

bandersnatch also has a short Weierstrass curve form, following the form: y² = x³ + A * x + B where

A = 10773120815616481058602537765553212789256758185246796157495669123169359657269 B = 29569587568322301171008055308580903175558631321415017492731745847794083609535

Script to transfer between different curves are available https://github.com/zhenfeizhang/bandersnatch/blob/main/bandersnatch/script/bandersnatch.sage